requirements.txt
praw
python-dotenv
pandas
tqdm

